QuiplashBackend/QuiplashCW2324/requirements.txt
azure-functions==1.17.0
azure-core==1.29.5
azure-cosmos==4.5.1
certifi==2023.7.22
chardet==5.2.0
charset-normalizer==3.3.0
idna==3.4
requests==2.31.0
six==1.16.0
tinydb==4.8.0
typing_extensions==4.8.0
urllib3==2.0.6

